v.1.5.5 - отдаю готовый фронт (background-image отображается), добавил CabinetController.java, перенес indexController.java
