+++ src/main/java/ru/coolteam/earnpocketmoney/controllers/IndexController.java
@@ -1,0 +1,15 @@
+package ru.coolteam.earnpocketmoney.controllers;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.*;
+
+@Controller
+@RequiredArgsConstructor
+public class IndexController {
+
+    @GetMapping("/")
+    public String index(){
+        return "index";
+    }
+}
