+++ src/main/resources/static/css/blocks/home.css
@@ -1,8 +1,6 @@
 .home {
     overflow: hidden;
-    /* background-image: url('/src/main/resources/static/content/img/fon_ny_1-01.svg'); */
-    background-image: url([[@{/content/img/fon_ny_1-01.svg}]]);
-
+    background-image: url(../../content/img/fon_ny_1-01.svg);
 }
 
 .home__content__row {
@@ -12,7 +10,6 @@
 }
 
 .left__content__element {
-
 }
 
 .right__content__element {
@@ -24,8 +21,7 @@
     min-width: 39.5%;
     height: 1000px;
     min-height: 100%;
-    /* background-image: url('/src/main/resources/static/content/img/fon_ny_2-01.svg'); */
-    background-image: url([[@{/content/img/fon_ny_2-01.svg}]]);
+    background-image: url(../../content/img/fon_ny_2-01.svg);
 
     border: 1px solid #828282; 
     box-sizing: border-box;
