+++ src/main/resources/static/family.html
@@ -34,7 +34,7 @@
                         <li><a href="#">Главная</a></li>
                         <li><a href="#" class= "link__active">Кабинет</a></li>
                         <li><a href="#">Семья</a></li>
-                        <li><a href="/app/api/v2/tasks/all">Задачи</a></li>
+                        <li><a href="/app/api/v2/tasks/create">Задачи</a></li>
                         <li><a href="#">Статистика</a></li>
                     </ul>
                 </nav>
