+++ src/main/resources/static/login.html
@@ -12,17 +12,13 @@
       <link href="css/blocks/reset.css" rel="stylesheet" />
       <link href="css/fonts/VAGRounded/stylesheet.css" rel="stylesheet" />
       
-
-      <!-- <link rel="stylesheet" data-th-href="@{/login.css}" media="screen" /> -->
-      <link rel="stylesheet" th:href="@{/login.css}" media="screen" />
+      <link th:href="@{/login.css}" rel="stylesheet" />
       <link th:href="@{/css/blocks/base.css}" rel="stylesheet" />
       <link th:href="@{/css/blocks/reset.css}" rel="stylesheet" />
-      <link data-th-href="@{/css/fonts/VAGRounded/stylesheet.css}" rel="stylesheet" />
+      <link th:href="@{/css/fonts/VAGRounded/stylesheet.css}" rel="stylesheet" />
   </head>
 
-  <!-- <body th:style="'background: url(/img/fon_ny_1-01.svg}'"> -->
-  <!-- <body th:src="@{/src/main/resources/static/content/img/fon_ny_1-01.svg}"> -->
-  <body th:src="@{/img/fon_ny_1-01.svg}">
+  <body>
   <!-- <body> -->
     <div class="login">
       <div class="container login__container">
@@ -45,27 +41,28 @@
                   <span th:text="${message}"></span>
 
                   <input type="text"
-                          th:field="*{login}"
-                          class="form-control"
-                          placeholder="login"
-                          autofocus="true"/>
+                    th:field="*{login}"
+                    class="form-control"
+                    placeholder="login"
+                    autofocus="true"/>
         
                   <input type="password"
-                          th:field="*{password}"
-                          class="form-control"
-                          placeholder="Password"/>
+                    th:field="*{password}"
+                    class="form-control"
+                    placeholder="Password"/>
     
                   <span class="has-error"
                     th:text="${error}">
                   </span>
         
                   <input type="hidden"
-                            name="${_csrf.parameterName}"
-                            value="${_csrf.token}"/>
+                    name="${_csrf.parameterName}"
+                    value="${_csrf.token}"/>
         
                   <button class="btn btn-lg btn-primary btn-block" type="submit">Войти</button>
                     
                   <h4 class="text-center"><a href="/register">Создать новый аккаунт</a></h4>
+               
                 </div>
 
               </form>
@@ -85,64 +82,3 @@
     </div>
   </body>
 </html>
-
-
-<!--<body>-->
-<!--  <div class="login">-->
-<!--    <div class="container login__container">-->
-
-<!--      <div class="login__wrapper">-->
-
-<!--        <div class="login__col">-->
-
-<!--          <form method="POST"-->
-<!--                class="form-signin"-->
-<!--                th:object="${userForm}"-->
-<!--                th:action="@{/auth}">-->
-
-<!--            <h2 class="form-heading">Вход</h2>-->
-<!--    -->
-<!--            <div class="form-group">-->
-<!--                <span th:text="${message}"></span>-->
-
-<!--                  &lt;!&ndash; <input type="text"&ndash;&gt;-->
-<!--                  &lt;!&ndash;        name="username"&ndash;&gt;-->
-<!--                  &lt;!&ndash;        class="form-control" &ndash;&gt;-->
-<!--                  &lt;!&ndash;        placeholder="Username" &ndash;&gt;-->
-<!--                  &lt;!&ndash;        autofocus="true"/>&ndash;&gt;-->
-
-<!--                <input type="text"-->
-<!--                       th:field="*{login}"-->
-<!--                       class="form-control"-->
-<!--                       placeholder="login"-->
-<!--                       autofocus="true"/>-->
-<!--    -->
-<!--                <input type="password"-->
-<!--                       th:field="*{password}"-->
-<!--                       class="form-control"-->
-<!--                       placeholder="Password"/>-->
-
-<!--                <span class="has-error"-->
-<!--                  th:text="${error}">-->
-<!--                </span>-->
-<!--    -->
-<!--                <input type="hidden"-->
-<!--                        name="${_csrf.parameterName}"-->
-<!--                        value="${_csrf.token}"/>-->
-<!--    -->
-<!--                <button class="btn btn-lg btn-primary btn-block" type="submit">Log In</button>-->
-<!--                -->
-<!--                <h4 class="text-center"><a href="/register">Create an account</a></h4>-->
-<!--            </div>-->
-<!--          </form>-->
-<!--        </div>-->
-<!--      </div>-->
-<!--      -->
-<!--    </div>-->
-
-<!--    <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js" integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj" crossorigin="anonymous"></script>-->
-<!--    <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.1/dist/umd/popper.min.js" integrity="sha384-9/reFTGAW83EW2RDu2S0VKaIzap3H66lZH81PoYlFhbGU+6BZp6G7niu735Sk7lN" crossorigin="anonymous"></script>-->
-<!--    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js" integrity="sha384-B4gt1jrGC7Jh4AgTPSdUtOBvfO8shuf57BaghqFfPlYxofvL8/KUEfYiJOMMV+rV" crossorigin="anonymous"></script>-->
-<!--      &lt;!&ndash;    <script src="/js/main.js"></script>&ndash;&gt;-->
-<!--  </div>-->
-<!--</body>-->
