+++ src/main/resources/static/registration.html
@@ -8,76 +8,86 @@
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.1.1/jquery.min.js"></script>
     <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js"></script>
 
-    <!-- <link href="/css/common.css" rel="stylesheet"> -->
-    <link href="/src/main/resources/static/css/blocks/registration.css" rel="stylesheet" >
+    <link href="css/blocks/registration.css" rel="stylesheet" >
 
     <link href="css/blocks/base.css" rel="stylesheet" />
     <link href="css/blocks/reset.css" rel="stylesheet" />
-    
     <link href="css/fonts/VAGRounded/stylesheet.css" rel="stylesheet" />
+
+    <link th:href="@{/registration.css}" rel="stylesheet"/>
+    <link th:href="@{/css/blocks/base.css}" rel="stylesheet" />
+    <link th:href="@{/css/blocks/reset.css}" rel="stylesheet" />
+    <link th:href="@{/css/fonts/VAGRounded/stylesheet.css}" rel="stylesheet" />
 </head>
 
 <body>
     <div class="registration">
-        <div class="registration__container">
+        <div class="container registration__container">
 
-            <div class="registration__wrapper">
+            <div class=" content registration__content">
         
-                <div class="registration__col">
-        
-                    <form method="POST"
+                <div class="content__element">
+
+                    <div class="board registration__board">
+
+                        <form method="POST"
                             class="form-signin"
                             th:object="${userForm}"
                             th:action="@{/register}">
-        
-                        <h2 class="form-signin-heading">Create your account</h2>
-        
-                        <div class="form-group">
+    
+                            <h2 class="header__h2 board__header__h2 registration__board__header">Create your account</h2>
+    
+                            <div class="form-group">
 
-                            <input type="text"
-                                   th:field="*{login}"
-                                   class="form-control"
-                                   placeholder="login"
-                                   autofocus="true">
-        
-                            <span class="has-error"
+                                <input type="text"
+                                    th:field="*{login}"
+                                    class="form-control"
+                                    placeholder="login"
+                                    autofocus="true">
+            
+                                <span class="has-error"
                                     th:if="${#fields.hasErrors('login')}"
                                     th:errors="*{login}">
-                            </span>
-        
-                        </div>
-                
-                        <div class="form-group">
-                            <input type="password"
-                                   th:field="*{password}"
-                                   class="form-control"
-                                   placeholder="Password">
-        
-                            <span class="has-error"
+                                </span>
+            
+                            </div>
+            
+                            <div class="form-group">
+                                <input type="password"
+                                    th:field="*{password}"
+                                    class="form-control"
+                                    placeholder="Password">
+            
+                                <span class="has-error"
                                     th:if="${#fields.hasErrors('password')}"
                                     th:errors="*{password}">
-                            </span>
-        
-                        </div>
-                
-                        <div class="form-group">
-                            <input type="password"
+                                </span>
+            
+                            </div>
+            
+                            <div class="form-group">
+                                <input type="password"
                                     th:field="*{passwordConfirm}"
                                     class="form-control"
                                     placeholder="Confirm your password">
+            
+                                <span class="has-error"
+                                    th:if="${#fields.hasErrors('passwordConfirm')}"
+                                    th:errors="*{passwordConfirm}">
+                                </span>
+            
+                            </div>
+            
+                            <button class="btn btn-lg btn-primary btn-block" type="submit">Submit</button>
+                    
+                        </form>
+
+                    </div>
         
-                            <span class="has-error"
-                                th:if="${#fields.hasErrors('passwordConfirm')}"
-                                th:errors="*{passwordConfirm}">
-                            </span>
-        
-                        </div>
-                
-                        <button class="btn btn-lg btn-primary btn-block" type="submit">Submit</button>
-                        
-                    </form>
                 </div>
-            </div>    
+
+            </div>  
+
         </div>
         
         <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js" integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj" crossorigin="anonymous"></script>
