+++ src/main/resources/static/tasks.html
@@ -6,11 +6,22 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Earn Pocket Money</title>
 
-    <link rel="stylesheet" href="css/blocks/header.css">
-    <!-- <link rel="stylesheet" href="css/blocks/home.css"> -->
+
+    <link th:href="@{css/blocks/board_task_create.css}"  media="screen" rel="stylesheet" >
+    <link th:href="@{css/blocks/board_tasks_list.css}"  media="screen" rel="stylesheet" >
+    <link th:href="@{css/blocks/board_tasks_examples.css}"  media="screen" rel="stylesheet" >
+
+    <link th:href="@{css/blocks/header.css}" media="screen" rel="stylesheet">
+    <link th:href="@{css/blocks/base.css}" media="screen" rel="stylesheet">
+    <link th:href="@{css/blocks/reset.css}" media="screen" rel="stylesheet">
+    <link th:href="@{/css/fonts/VAGRounded/stylesheet.css}" rel="stylesheet" />
+
+
     <link rel="stylesheet" href="css/blocks/board_task_create.css">
-    <!-- <link rel="stylesheet" href="css/blocks/board_tasks_list.css"> -->
+    <link rel="stylesheet" href="css/blocks/board_tasks_list.css">
     <link rel="stylesheet" href="css/blocks/board_tasks_examples.css">
+
+    <link rel="stylesheet" href="css/blocks/header.css">
 
     <!--Reset-->
     <link rel="stylesheet" href="css/blocks/base.css">
@@ -25,13 +36,15 @@
     <header class="header">
         <div class="container">
             <div class="header-row">
-                <div class="header-logo"><img src="content/img/logo-01.svg" alt="Logo" width="250px"></div>
+                <div class="header-logo">
+                    <img src="content/img/logo-01.svg" alt="Logo" width="250px" th:src="@{/content/img/logo-01.svg}">
+                </div>
                 <nav class="header-nav">
                     <ul class="">
                         <li><a href="#">Главная</a></li>
-                        <li><a href="#" class= "link__active">Кабинет</a></li>
+                        <li><a th:href="@{/api/v1/cabinet}">Кабинет</a></li>
                         <li><a href="#">Семья</a></li>
-                        <li><a href="/app/api/v2/tasks/all">Задачи</a></li>
+                        <li><a th:href="@{/api/v1/tasks/all}" class= "link__active">Задачи</a></li>
                         <li><a href="#">Статистика</a></li>
                     </ul>
                 </nav>
